Bump pymongo to 3.8.0 to fix auth issues
See https://api.mongodb.com/python/current/changelog.html#changes-in-version-3-7-1

Since KBase is currently on MongoDB 3.6 with MMAPv1, updating to pymongo
3.9.0 is not safe without further changes. Ideally we update to WiredTiger
soon as MMAPv1 is deprecated in Mongo 4.0 and removed in 4.2. See
https://api.mongodb.com/python/current/changelog.html#changes-in-version-3-9-0

Alternatively we can add a config option to turn off retryable writes if
we need to upgrade to 3.9.0 and WiredTiger isn't deployed yet.